lxd/network/acl: Read OVN logs from systemd journal
In the case of an OVN controller being deployed as part of a MicroOVN deployment,
the OVN controller logs are stored in MicroOVN's snap syslog. The LXD snap should have root access,
which means that it should be authorized (this is being tested) to read the OVN controller logs.

In the case where logs couldn't be read (e.g, AppArmor restrictions), we could try to connect the `log-observe` snap interface on LXD
and see if this works.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>