lxd/network/acl: If the OVN chassis is in MicroOVN, read the ACL logs from syslogs
If the OVN controller is deployed as part of MicroOVN, it means that the `ovn-chassis` snap connection hook
between LXD and MicroOVN has been fired and that a symlink like:

`/run/openvswitch -> /var/snap/microovn/...`

exists. Otherwise, there might still be a symlink but not with that target prefix. That's how we detect the ovn chassis
connection without introducing a new plug / slot between LXD and MicroOVN.

Then, if this case is detected, we check the status of the systemd unit 'snap.microovn.chassis.service' (the one containing the OVN controller in its sd journal)
If it is loaded and active, we read the last 5000 ACL log entries (we can discuss on this limit) in the journal and return them.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>